+++ components/modal/PurePanel.tsx
@@ -11,7 +11,6 @@ import { ConfirmContent } from './ConfirmDialog';
 import type { ModalFuncProps } from './interface';
 import { Footer, renderCloseIcon } from './shared';
 import useStyle from './style';
-import WireframeCmp from './style/wireframeCmp';
 
 export interface PurePanelProps
   extends Omit<PanelProps, 'prefixCls'>,
@@ -67,23 +66,20 @@ const PurePanel: React.FC<PurePanelProps> = (props) => {
   }
 
   return (
-    <>
-      <Panel
-        prefixCls={prefixCls}
-        className={classNames(
-          hashId,
-          `${prefixCls}-pure-panel`,
-          type && confirmPrefixCls,
-          type && `${confirmPrefixCls}-${type}`,
-          className,
-        )}
-        {...restProps}
-        closeIcon={renderCloseIcon(prefixCls, closeIcon)}
-        closable={closable}
-        {...additionalProps}
-      />
-      <WireframeCmp prefixCls={prefixCls} />
-    </>
+    <Panel
+      prefixCls={prefixCls}
+      className={classNames(
+        hashId,
+        `${prefixCls}-pure-panel`,
+        type && confirmPrefixCls,
+        type && `${confirmPrefixCls}-${type}`,
+        className,
+      )}
+      {...restProps}
+      closeIcon={renderCloseIcon(prefixCls, closeIcon)}
+      closable={closable}
+      {...additionalProps}
+    />
   );
 };
